api: factor shared JSON Schema builder out of the OpenAPI generator
argsObjectSchema()/ARG_TYPE_MAP move from generate-api-docs.mts into
src/lib/api/json-schema.ts so the upcoming MCP manifest generator can
reuse the identical arg->JSON-Schema derivation instead of a second copy
(R-3.1) — REST and MCP tool input schemas can't drift from each other.
Output-identical: pnpm run api:docs:check passes unchanged.
